+++ src/index.ts
@@ -1,1 +1,12 @@
-export { PubSub, SubscriptionManager } from './pubsub';
+export { PubSubEngine } from './pubsub-engine';
+export { PubSub } from './pubsub';
+export { withFilter, ResolverFn, FilterFn } from './with-filter';
+export {
+  SubscriptionManager,
+  SubscriptionOptions,
+  TriggerConfig,
+  TriggerMap,
+  SetupFunction,
+  SetupFunctions,
+  ValidationError,
+} from './subscriptions-manager';